requirements.txt
# Python 3.11 Compatible Requirements - RECOMMENDED
# Generated for media processing automation tool

# Core Web Framework
Flask==3.1.1
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0

# File System Monitoring
watchdog==3.0.0

# AI/ML Libraries
openai==1.51.0
pyannote.audio==3.3.2
pytorch-lightning==2.5.0
torchmetrics==1.4.0

# Audio/Video Processing
opencv-python==4.10.0.82
ffmpeg-python==0.2.0
moviepy==1.0.3
scenedetect[opencv]==0.6.6

# Scientific Computing
numpy==1.26.4
scipy==1.11.4

# Additional Dependencies (install if needed)
# torch>=2.0.0
# torchaudio>=2.0.0
# torchvision>=0.15.0
# transformers>=4.30.0
# librosa>=0.10.0
# soundfile>=0.12.0
# matplotlib>=3.7.0
# pandas>=2.0.0
